+++ cocos/physics/physx/physx-world.ts
@@ -31,218 +31,22 @@
 /* eslint-disable @typescript-eslint/no-unsafe-return */
 import { Ray } from '../../core/geometry';
 import { IPhysicsWorld, IRaycastOptions } from '../spec/i-physics-world';
-import { CollisionEventType, PhysicsMaterial, PhysicsRayResult, TriggerEventType } from '../framework';
+import { PhysicsMaterial, PhysicsRayResult, CollisionEventType, TriggerEventType } from '../framework';
 import { error, Node, RecyclePool } from '../../core';
 import { IVec3Like } from '../../core/math/type-define';
 import { IBaseConstraint } from '../spec/i-physics-constraint';
-import { PhysXSharedBody } from './physx-shared-body';
 import { PhysXRigidBody } from './physx-rigid-body';
-import { PhysXShape } from './shapes/physx-shape';
-import { PhysXContactEquation } from './physx-contact-equation';
-import { CollisionEventObject, TriggerEventObject } from '../utils/util';
-import { addActorToScene, getContactData, raycastAll, getWrapShape, PX, simulateScene,
-    initializeWorld, raycastClosest, EFilterDataWord3 } from './export-physx';
+import {
+    addActorToScene, raycastAll, simulateScene, initializeWorld, raycastClosest,
+    gatherEvents, getWrapShape, PX, getContactDataOrByteOffset,
+} from './export-physx';
+import { PhysXSharedBody } from './physx-shared-body';
 import { fastRemoveAt } from '../../core/utils/array';
 import { TupleDictionary } from '../utils/tuple-dictionary';
-
-interface ITriggerEventItem {
-    a: PhysXShape,
-    b: PhysXShape,
-    times: number,
-}
-
-interface ICollisionEventItem {
-    type: CollisionEventType,
-    a: PhysXShape,
-    b: PhysXShape,
-    contactCount: number,
-    buffer: any,
-    offset: number,
-}
-
-const triggerEventBeginDic = new TupleDictionary();
-const triggerEventEndDic = new TupleDictionary();
-const triggerEventsPool: ITriggerEventItem[] = [];
-function onTrigger (type: TriggerEventType, wpa: PhysXShape, wpb: PhysXShape): void {
-    if (wpa && wpb) {
-        if (wpa.collider.needTriggerEvent || wpb.collider.needTriggerEvent) {
-            let tE: ITriggerEventItem;
-            if (triggerEventsPool.length > 0) {
-                tE = triggerEventsPool.pop() as ITriggerEventItem;
-                tE.a = wpa; tE.b = wpb; tE.times = 0;
-            } else {
-                tE = { a: wpa, b: wpb, times: 0 };
-            }
-            if (type === 'onTriggerEnter') {
-                triggerEventBeginDic.set(wpa.id, wpb.id, tE);
-            } else {
-                triggerEventEndDic.set(wpa.id, wpb.id, tE);
-            }
-        }
-    }
-}
-
-function emitTriggerEvent () {
-    let len = triggerEventEndDic.getLength();
-    while (len--) {
-        const key = triggerEventEndDic.getKeyByIndex(len);
-        const data = triggerEventEndDic.getDataByKey<ITriggerEventItem>(key);
-        triggerEventsPool.push(data);
-        const dataBeg = triggerEventBeginDic.getDataByKey<ITriggerEventItem>(key);
-        if (dataBeg) {
-            triggerEventsPool.push(dataBeg);
-            triggerEventBeginDic.set(data.a.id, data.b.id, null);
-        }
-        const colliderA = data.a.collider;
-        const colliderB = data.b.collider;
-        if (colliderA && colliderB) {
-            const type: TriggerEventType = 'onTriggerExit';
-            TriggerEventObject.type = type;
-            if (colliderA.needTriggerEvent) {
-                TriggerEventObject.selfCollider = colliderA;
-                TriggerEventObject.otherCollider = colliderB;
-                colliderA.emit(type, TriggerEventObject);
-            }
-            if (colliderB.needTriggerEvent) {
-                TriggerEventObject.selfCollider = colliderB;
-                TriggerEventObject.otherCollider = colliderA;
-                colliderB.emit(type, TriggerEventObject);
-            }
-        }
-    }
-    triggerEventEndDic.reset();
-
-    len = triggerEventBeginDic.getLength();
-    while (len--) {
-        const key = triggerEventBeginDic.getKeyByIndex(len);
-        const data = triggerEventBeginDic.getDataByKey<ITriggerEventItem>(key);
-        const colliderA = data.a.collider;
-        const colliderB = data.b.collider;
-        if (!colliderA || !colliderA.isValid || !colliderB || !colliderB.isValid) {
-            triggerEventsPool.push(data);
-            triggerEventBeginDic.set(data.a.id, data.b.id, null);
-        } else {
-            const type: TriggerEventType = data.times++ ? 'onTriggerStay' : 'onTriggerEnter';
-            TriggerEventObject.type = type;
-            if (colliderA.needTriggerEvent) {
-                TriggerEventObject.selfCollider = colliderA;
-                TriggerEventObject.otherCollider = colliderB;
-                colliderA.emit(type, TriggerEventObject);
-            }
-            if (colliderB.needTriggerEvent) {
-                TriggerEventObject.selfCollider = colliderB;
-                TriggerEventObject.otherCollider = colliderA;
-                colliderB.emit(type, TriggerEventObject);
-            }
-        }
-    }
-}
-
-const contactEventDic = new TupleDictionary();
-const contactEventsPool: ICollisionEventItem[] = [];
-/**
- * @param c the contact count, how many the contacts in this pair
- * @param d the contact buffer, the buffer stores all contacts
- * @param o the data offset, the first contact offset in the contact buffer
- */
-function onCollision (type: CollisionEventType, wpa: PhysXShape, wpb: PhysXShape, c: number, d: any, o: number): void {
-    if (wpa && wpb) {
-        if (wpa.collider.needCollisionEvent || wpb.collider.needCollisionEvent) {
-            if (contactEventsPool.length > 0) {
-                const cE = contactEventsPool.pop() as ICollisionEventItem;
-                cE.type = type; cE.a = wpa; cE.b = wpb; cE.contactCount = c; cE.buffer = d; cE.offset = o;
-                contactEventDic.set(wpa.id, wpb.id, cE);
-            } else {
-                const cE: ICollisionEventItem = { type, a: wpa, b: wpb, contactCount: c, buffer: d, offset: o };
-                contactEventDic.set(wpa.id, wpb.id, cE);
-            }
-        }
-    }
-}
-
-const contactsPool: [] = [];
-function emitCollisionEvent (): void {
-    let len = contactEventDic.getLength();
-    while (len--) {
-        const key = contactEventDic.getKeyByIndex(len);
-        const data = contactEventDic.getDataByKey<ICollisionEventItem>(key);
-        contactEventsPool.push(data);
-        const colliderA = data.a.collider;
-        const colliderB = data.b.collider;
-        if (colliderA && colliderA.isValid && colliderB && colliderB.isValid) {
-            CollisionEventObject.type = data.type;
-            CollisionEventObject.impl = data.buffer;
-            const c = data.contactCount; const d = data.buffer; const o = data.offset;
-            const contacts = CollisionEventObject.contacts;
-            contactsPool.push.apply(contactsPool, contacts as any);
-            contacts.length = 0;
-            for (let i = 0; i < c; i++) {
-                if (contactsPool.length > 0) {
-                    const c = contactsPool.pop() as unknown as PhysXContactEquation;
-                    c.colliderA = colliderA; c.colliderB = colliderB;
-                    c.impl = getContactData(d, i, o); contacts.push(c);
-                } else {
-                    const c = new PhysXContactEquation(CollisionEventObject);
-                    c.colliderA = colliderA; c.colliderB = colliderB;
-                    c.impl = getContactData(d, i, o); contacts.push(c);
-                }
-            }
-            if (colliderA.needCollisionEvent) {
-                CollisionEventObject.selfCollider = colliderA;
-                CollisionEventObject.otherCollider = colliderB;
-                colliderA.emit(CollisionEventObject.type, CollisionEventObject);
-            }
-            if (colliderB.needCollisionEvent) {
-                CollisionEventObject.selfCollider = colliderB;
-                CollisionEventObject.otherCollider = colliderA;
-                colliderB.emit(CollisionEventObject.type, CollisionEventObject);
-            }
-        }
-    }
-    contactEventDic.reset();
-}
-
-const eventCallback = {
-    onContactBegin: (a: any, b: any, c: any, d: any, o: number): void => {
-        const wpa = getWrapShape<PhysXShape>(a);
-        const wpb = getWrapShape<PhysXShape>(b);
-        onCollision('onCollisionEnter', wpa, wpb, c, d, o);
-    },
-    onContactEnd: (a: any, b: any, c: any, d: any, o: number): void => {
-        const wpa = getWrapShape<PhysXShape>(a);
-        const wpb = getWrapShape<PhysXShape>(b);
-        onCollision('onCollisionExit', wpa, wpb, c, d, o);
-    },
-    onContactPersist: (a: any, b: any, c: any, d: any, o: number): void => {
-        const wpa = getWrapShape<PhysXShape>(a);
-        const wpb = getWrapShape<PhysXShape>(b);
-        onCollision('onCollisionStay', wpa, wpb, c, d, o);
-    },
-    onTriggerBegin: (a: any, b: any): void => {
-        const wpa = getWrapShape<PhysXShape>(a);
-        const wpb = getWrapShape<PhysXShape>(b);
-        onTrigger('onTriggerEnter', wpa, wpb);
-    },
-    onTriggerEnd: (a: any, b: any): void => {
-        const wpa = getWrapShape<PhysXShape>(a);
-        const wpb = getWrapShape<PhysXShape>(b);
-        onTrigger('onTriggerExit', wpa, wpb);
-    },
-};
-
-// eNONE = 0,   //!< the query should ignore this shape
-// eTOUCH = 1,  //!< a hit on the shape touches the intersection geometry of the query but does not block it
-// eBLOCK = 2   //!< a hit on the shape blocks the query (does not block overlap queries)
-const queryCallback = {
-    preFilter (filterData: any, shape: any, _actor: any, _out: any): number {
-        const shapeFlags = shape.getFlags();
-        if ((filterData.word3 & EFilterDataWord3.QUERY_CHECK_TRIGGER)
-            && (shapeFlags & PX.ShapeFlag.eTRIGGER_SHAPE)) {
-            return PX.QueryHitType.eNONE;
-        }
-        return filterData.word3 & EFilterDataWord3.QUERY_SINGLE_HIT ? PX.QueryHitType.eBLOCK : PX.QueryHitType.eTOUCH;
-    },
-};
+import { PhysXContactEquation } from './physx-contact-equation';
+import { CollisionEventObject, TriggerEventObject } from '../utils/util';
+import { PhysXShape } from './shapes/physx-shape';
+import { EFilterDataWord3 } from './physx-enum';
 
 export class PhysXWorld implements IPhysicsWorld {
     setAllowSleep (_v: boolean): void { }
@@ -264,11 +68,12 @@ export class PhysXWorld implements IPhysicsWorld {
     readonly queryFilterCB: any;
 
     readonly wrappedBodies: PhysXSharedBody[] = [];
+    readonly callback = PhysXCallback;
 
     mutipleResultSize = 12;
 
     constructor () {
-        initializeWorld(this, eventCallback, queryCallback, onCollision, onTrigger);
+        initializeWorld(this);
     }
 
     destroy (): void {
@@ -336,7 +141,213 @@ export class PhysXWorld implements IPhysicsWorld {
     }
 
     emitEvents (): void {
-        emitTriggerEvent();
-        emitCollisionEvent();
+        gatherEvents(this);
+        PhysXCallback.emitTriggerEvent();
+        PhysXCallback.emitCollisionEvent();
     }
 }
+
+///
+/// Event Callback
+///
+
+interface ITriggerEventItem {
+    a: PhysXShape,
+    b: PhysXShape,
+    times: number,
+}
+
+interface ICollisionEventItem {
+    type: CollisionEventType,
+    a: PhysXShape,
+    b: PhysXShape,
+    contactCount: number,
+    buffer: any,
+    offset: number,
+}
+
+const triggerEventBeginDic = new TupleDictionary();
+const triggerEventEndDic = new TupleDictionary();
+const triggerEventsPool: ITriggerEventItem[] = [];
+const contactEventDic = new TupleDictionary();
+const contactEventsPool: ICollisionEventItem[] = [];
+const contactsPool: [] = [];
+const PhysXCallback = {
+    eventCallback: {
+        onContactBegin: (a: any, b: any, c: any, d: any, o: number): void => {
+            const wpa = getWrapShape<PhysXShape>(a);
+            const wpb = getWrapShape<PhysXShape>(b);
+            PhysXCallback.onCollision('onCollisionEnter', wpa, wpb, c, d, o);
+        },
+        onContactEnd: (a: any, b: any, c: any, d: any, o: number): void => {
+            const wpa = getWrapShape<PhysXShape>(a);
+            const wpb = getWrapShape<PhysXShape>(b);
+            PhysXCallback.onCollision('onCollisionExit', wpa, wpb, c, d, o);
+        },
+        onContactPersist: (a: any, b: any, c: any, d: any, o: number): void => {
+            const wpa = getWrapShape<PhysXShape>(a);
+            const wpb = getWrapShape<PhysXShape>(b);
+            PhysXCallback.onCollision('onCollisionStay', wpa, wpb, c, d, o);
+        },
+        onTriggerBegin: (a: any, b: any): void => {
+            const wpa = getWrapShape<PhysXShape>(a);
+            const wpb = getWrapShape<PhysXShape>(b);
+            PhysXCallback.onTrigger('onTriggerEnter', wpa, wpb, true);
+        },
+        onTriggerEnd: (a: any, b: any): void => {
+            const wpa = getWrapShape<PhysXShape>(a);
+            const wpb = getWrapShape<PhysXShape>(b);
+            PhysXCallback.onTrigger('onTriggerExit', wpa, wpb, false);
+        },
+    },
+
+    // eNONE = 0,   //!< the query should ignore this shape
+    // eTOUCH = 1,  //!< a hit on the shape touches the intersection geometry of the query but does not block it
+    // eBLOCK = 2   //!< a hit on the shape blocks the query (does not block overlap queries)
+    queryCallback: {
+        preFilter (filterData: any, shape: any, _actor: any, _out: any): number {
+            const shapeFlags = shape.getFlags();
+            if ((filterData.word3 & EFilterDataWord3.QUERY_CHECK_TRIGGER)
+                && (shapeFlags & PX.ShapeFlag.eTRIGGER_SHAPE)) {
+                return PX.QueryHitType.eNONE;
+            }
+            return filterData.word3 & EFilterDataWord3.QUERY_SINGLE_HIT ? PX.QueryHitType.eBLOCK : PX.QueryHitType.eTOUCH;
+        },
+    },
+
+    onTrigger (type: TriggerEventType, wpa: PhysXShape, wpb: PhysXShape, isEnter: boolean): void {
+        if (wpa && wpb) {
+            if (wpa.collider.needTriggerEvent || wpb.collider.needTriggerEvent) {
+                let tE: ITriggerEventItem;
+                if (triggerEventsPool.length > 0) {
+                    tE = triggerEventsPool.pop() as ITriggerEventItem;
+                    tE.a = wpa; tE.b = wpb; tE.times = 0;
+                } else {
+                    tE = { a: wpa, b: wpb, times: 0 };
+                }
+                if (isEnter) {
+                    triggerEventBeginDic.set(wpa.id, wpb.id, tE);
+                } else {
+                    triggerEventEndDic.set(wpa.id, wpb.id, tE);
+                }
+            }
+        }
+    },
+
+    emitTriggerEvent () {
+        let len = triggerEventEndDic.getLength();
+        while (len--) {
+            const key = triggerEventEndDic.getKeyByIndex(len);
+            const data = triggerEventEndDic.getDataByKey<ITriggerEventItem>(key);
+            triggerEventsPool.push(data);
+            const dataBeg = triggerEventBeginDic.getDataByKey<ITriggerEventItem>(key);
+            if (dataBeg) {
+                triggerEventsPool.push(dataBeg);
+                triggerEventBeginDic.set(data.a.id, data.b.id, null);
+            }
+            const colliderA = data.a.collider;
+            const colliderB = data.b.collider;
+            if (colliderA && colliderB) {
+                const type: TriggerEventType = 'onTriggerExit';
+                TriggerEventObject.type = type;
+                if (colliderA.needTriggerEvent) {
+                    TriggerEventObject.selfCollider = colliderA;
+                    TriggerEventObject.otherCollider = colliderB;
+                    colliderA.emit(type, TriggerEventObject);
+                }
+                if (colliderB.needTriggerEvent) {
+                    TriggerEventObject.selfCollider = colliderB;
+                    TriggerEventObject.otherCollider = colliderA;
+                    colliderB.emit(type, TriggerEventObject);
+                }
+            }
+        }
+        triggerEventEndDic.reset();
+
+        len = triggerEventBeginDic.getLength();
+        while (len--) {
+            const key = triggerEventBeginDic.getKeyByIndex(len);
+            const data = triggerEventBeginDic.getDataByKey<ITriggerEventItem>(key);
+            const colliderA = data.a.collider;
+            const colliderB = data.b.collider;
+            if (!colliderA || !colliderA.isValid || !colliderB || !colliderB.isValid) {
+                triggerEventsPool.push(data);
+                triggerEventBeginDic.set(data.a.id, data.b.id, null);
+            } else {
+                const type: TriggerEventType = data.times++ ? 'onTriggerStay' : 'onTriggerEnter';
+                TriggerEventObject.type = type;
+                if (colliderA.needTriggerEvent) {
+                    TriggerEventObject.selfCollider = colliderA;
+                    TriggerEventObject.otherCollider = colliderB;
+                    colliderA.emit(type, TriggerEventObject);
+                }
+                if (colliderB.needTriggerEvent) {
+                    TriggerEventObject.selfCollider = colliderB;
+                    TriggerEventObject.otherCollider = colliderA;
+                    colliderB.emit(type, TriggerEventObject);
+                }
+            }
+        }
+    },
+
+    /**
+     * @param c the contact count, how many the contacts in this pair
+     * @param d the contact buffer, the buffer stores all contacts
+     * @param o the data offset, the first contact offset in the contact buffer
+     */
+    onCollision (type: CollisionEventType, wpa: PhysXShape, wpb: PhysXShape, c: number, d: any, o: number): void {
+        if (wpa && wpb) {
+            if (wpa.collider.needCollisionEvent || wpb.collider.needCollisionEvent) {
+                if (contactEventsPool.length > 0) {
+                    const cE = contactEventsPool.pop() as ICollisionEventItem;
+                    cE.type = type; cE.a = wpa; cE.b = wpb; cE.contactCount = c; cE.buffer = d; cE.offset = o;
+                    contactEventDic.set(wpa.id, wpb.id, cE);
+                } else {
+                    const cE: ICollisionEventItem = { type, a: wpa, b: wpb, contactCount: c, buffer: d, offset: o };
+                    contactEventDic.set(wpa.id, wpb.id, cE);
+                }
+            }
+        }
+    },
+
+    emitCollisionEvent (): void {
+        let len = contactEventDic.getLength();
+        while (len--) {
+            const key = contactEventDic.getKeyByIndex(len);
+            const data = contactEventDic.getDataByKey<ICollisionEventItem>(key);
+            contactEventsPool.push(data);
+            const colliderA = data.a.collider;
+            const colliderB = data.b.collider;
+            if (colliderA && colliderA.isValid && colliderB && colliderB.isValid) {
+                CollisionEventObject.type = data.type;
+                CollisionEventObject.impl = data.buffer;
+                const c = data.contactCount; const d = data.buffer; const o = data.offset;
+                const contacts = CollisionEventObject.contacts;
+                contactsPool.push.apply(contactsPool, contacts as any);
+                contacts.length = 0;
+                for (let i = 0; i < c; i++) {
+                    if (contactsPool.length > 0) {
+                        const c = contactsPool.pop() as unknown as PhysXContactEquation;
+                        c.colliderA = colliderA; c.colliderB = colliderB;
+                        c.impl = getContactDataOrByteOffset(i, o); contacts.push(c);
+                    } else {
+                        const c = new PhysXContactEquation(CollisionEventObject);
+                        c.colliderA = colliderA; c.colliderB = colliderB;
+                        c.impl = getContactDataOrByteOffset(i, o); contacts.push(c);
+                    }
+                }
+                if (colliderA.needCollisionEvent) {
+                    CollisionEventObject.selfCollider = colliderA;
+                    CollisionEventObject.otherCollider = colliderB;
+                    colliderA.emit(CollisionEventObject.type, CollisionEventObject);
+                }
+                if (colliderB.needCollisionEvent) {
+                    CollisionEventObject.selfCollider = colliderB;
+                    CollisionEventObject.otherCollider = colliderA;
+                    colliderB.emit(CollisionEventObject.type, CollisionEventObject);
+                }
+            }
+        }
+        contactEventDic.reset();
+    },
+};
