+++ cocos/physics/physx/shapes/physx-shape.ts
@@ -33,7 +33,9 @@ import { IVec3Like, Quat, Vec3 } from '../../../core';
 import { AABB, Sphere } from '../../../core/geometry';
 import { Collider, RigidBody, PhysicsMaterial, PhysicsSystem } from '../../framework';
 import { IBaseShape } from '../../spec/i-physics-shape';
-import { EFilterDataWord3, getShapeFlags, getShapeMaterials, getShapeWorldBounds, getTempTransform, PX, _pxtrans, _trans } from '../export-physx';
+import { addReference, getShapeFlags, getShapeMaterials, getShapeWorldBounds, getTempTransform,
+    PX, removeReference, _pxtrans, _trans } from '../export-physx';
+import { EFilterDataWord3 } from '../physx-enum';
 import { PhysXSharedBody } from '../physx-shared-body';
 import { PhysXWorld } from '../physx-world';
 
@@ -83,13 +85,7 @@ export class PhysXShape implements IBaseShape {
         this._sharedBody = (PhysicsSystem.instance.physicsWorld as PhysXWorld).getSharedBody(v.node);
         this._sharedBody.reference = true;
         this.onComponentSet();
-        if (this._impl) {
-            if (this._impl.$$) {
-                PX.IMPL_PTR[this._impl.$$.ptr] = this;
-            } else {
-                PX.IMPL_PTR[this.id] = this;
-            }
-        }
+        addReference(this, this._impl);
     }
 
     setIndex (v: number): void {
@@ -119,13 +115,7 @@ export class PhysXShape implements IBaseShape {
 
     onDestroy (): void {
         this._sharedBody.reference = false;
-        if (this._impl.$$) {
-            PX.IMPL_PTR[this._impl.$$.ptr] = null;
-            delete PX.IMPL_PTR[this._impl.$$.ptr];
-        } else {
-            PX.IMPL_PTR[this.id] = null;
-            delete PX.IMPL_PTR[this.id];
-        }
+        removeReference(this, this._impl);
         this._impl.release();
         this._impl = null;
     }
